requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
croniter==0.3.37
graia-application-mirai==0.18.4
graia-broadcast==0.8.11
idna==3.2
iterwrapper==0.1.4
multidict==5.1.0
natsort==7.1.1
pydantic==1.7.1
python-dateutil==2.8.2
regex==2020.11.13
requests==2.26.0
selenium==2.48.0
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
